# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-Sign Workflow API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createAdminUser
      tags: [admin]
      summary: Create a user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "200":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/audit:
    get:
      operationId: getAdminAuditLog
      tags: [admin]
      summary: Get audit log of all document and signing events (admin only)
      responses:
        "200":
          description: Audit log events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}/role:
    patch:
      operationId: updateAdminUserRole
      tags: [admin]
      summary: Update a user's role (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, user, auditor]
        provider:
          type: string
          enum: [local, azure]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - provider
        - createdAt
    AdminUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
      required:
        - users
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        role:
          type: string
          enum: [admin, user, auditor]
      required:
        - name
        - email
        - password
    UpdateRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [admin, user, auditor]
      required:
        - role
    AuditEvent:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [uploaded, sent, viewed, signed, completed]
        documentId:
          type: string
        documentTitle:
          type: string
        actorName:
          type: string
          nullable: true
        actorEmail:
          type: string
          nullable: true
        ipAddress:
          type: string
          nullable: true
        timestamp:
          type: string
      required:
        - id
        - type
        - documentId
        - documentTitle
        - actorName
        - actorEmail
        - ipAddress
        - timestamp
    AuditLogResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/AuditEvent"
      required:
        - events
...